Separate classes in single-class file #14
index file moved outside of src folder
AppSettings.TIME_DEBUG_PARAMS
