+++ app/lib/game/index.ts
@@ -1,0 +1,2 @@
+export * from './src/game';
+export * from './src/gameObject';
