--- app/lib/game/src/index.ts
@@ -1,2 +1,0 @@
-export * from './game';
-export * from './gameObject';
