+++ app/lib/graphics/src/types.ts
@@ -1,4 +1,4 @@
-import { Color } from '.';
+import { Color } from '..';
 
 export type SubStyle = Color | CanvasGradient | CanvasPattern;
