+++ app/lib/inputs/index.ts
@@ -1,0 +1,6 @@
+export * from './src/button';
+export * from './src/input';
+export * from './src/inputTimer';
+export * from './src/mouse';
+export * from './src/types';
+export * from './src/wheelAxis';
