--- app/lib/inputs/src/index.ts
@@ -1,6 +1,0 @@
-export * from './button';
-export * from './input';
-export * from './inputTimer';
-export * from './mouse';
-export * from './types';
-export * from './wheelAxis';
