+++ app/lib/rigid/src/bodies/rigidBody.ts
@@ -1,7 +1,7 @@
 import { GameCycle } from '@gandolphinnn/shared';
 import { CnvElement } from '@gandolphinnn/graphics'
 import { GameObject } from '@gandolphinnn/game'
-import { Collision, ERigidBodyEvent, LayerMask, Vector } from '..';
+import { Collision, ERigidBodyEvent, LayerMask, Vector } from '../..';
 
 export abstract class RigidBody implements GameCycle {
