+++ tsconfig.json
@@ -10,10 +10,10 @@
 		"baseUrl": "./",
 		"paths": {
 			"@gandolphinnn/shared": ["app/lib/shared/index.ts"],
-			"@gandolphinnn/graphics": ["app/lib/graphics/src/index.ts"],
-			"@gandolphinnn/inputs": ["app/lib/inputs/src/index.ts"],
-			"@gandolphinnn/rigid": ["app/lib/rigid/src/index.ts"],
-			"@gandolphinnn/game": ["app/lib/game/src/index.ts"],
+			"@gandolphinnn/graphics": ["app/lib/graphics/index.ts"],
+			"@gandolphinnn/inputs": ["app/lib/inputs/index.ts"],
+			"@gandolphinnn/rigid": ["app/lib/rigid/index.ts"],
+			"@gandolphinnn/game": ["app/lib/game/index.ts"],
 			"*": ["node_modules"]
 		}
 	},
